Move all final stage SQL builders to a single file for cleaner organization
