+++ datajunction-server/datajunction_server/api/sql.py
@@ -2,11 +2,9 @@
 SQL related APIs.
 """
 
-import json
 import logging
-from collections import OrderedDict
 from http import HTTPStatus
-from typing import Any, List, Optional, Tuple, cast
+from typing import List, Optional
 
 from fastapi import BackgroundTasks, Depends, Query, Request
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -17,27 +15,18 @@ from datajunction_server.internal.caching.query_cache_manager import (
     QueryCacheManager,
     QueryRequestParams,
 )
-from datajunction_server.api.helpers import (
-    assemble_column_metadata,
-    build_sql_for_multiple_metrics,
-    get_query,
-    validate_orderby,
-)
 from datajunction_server.internal.caching.cachelib_cache import get_cache
 from datajunction_server.internal.caching.interface import Cache
-from datajunction_server.database import Engine, Node
-from datajunction_server.database.queryrequest import QueryBuildType, QueryRequest
+from datajunction_server.database import Node
+from datajunction_server.database.queryrequest import QueryBuildType
 from datajunction_server.database.user import User
 from datajunction_server.errors import DJInvalidInputException
 from datajunction_server.internal.access.authentication.http import SecureAPIRouter
 from datajunction_server.internal.access.authorization import validate_access
-from datajunction_server.internal.engines import get_engine
 from datajunction_server.models import access
-from datajunction_server.models.access import AccessControlStore
 from datajunction_server.models.metric import TranslatedSQL
 from datajunction_server.models.node_type import NodeType
 from datajunction_server.models.sql import GeneratedSQL
-from datajunction_server.models.user import UserOutput
 from datajunction_server.utils import (
     get_current_user,
     get_session,
@@ -122,161 +111,6 @@ async def get_measures_sql_for_cube_v2(
     )
 
 
-async def build_node_sql(
-    node_name: str,
-    dimensions: List[str] = Query([]),
-    filters: List[str] = Query([]),
-    orderby: List[str] = Query([]),
-    limit: Optional[int] = None,
-    *,
-    session: AsyncSession = Depends(get_session),
-    engine: Engine | None = None,
-    access_control: AccessControlStore | None = None,
-    ignore_errors: bool = True,
-    use_materialized: bool = True,
-    query_parameters: dict[str, Any] | None = None,
-) -> TranslatedSQL:
-    """
-    Build node SQL and save it to query requests
-    """
-    validate_orderby(orderby, [node_name], dimensions)
-
-    node = cast(
-        Node,
-        await Node.get_by_name(session, node_name, raise_if_not_exists=True),
-    )
-    if not engine:
-        engine = node.current.catalog.engines[0]
-
-    # If it's a cube, we'll build SQL for the metrics in the cube, along with any additional
-    # dimensions or filters provided in the arguments
-    if node.type == NodeType.CUBE:
-        node = cast(
-            Node,
-            await Node.get_cube_by_name(session, node_name),
-        )
-        dimensions = list(
-            OrderedDict.fromkeys(node.current.cube_node_dimensions + dimensions),
-        )
-        translated_sql, engine, _ = await build_sql_for_multiple_metrics(
-            session=session,
-            metrics=node.current.cube_node_metrics,
-            dimensions=dimensions,
-            filters=filters,
-            orderby=orderby,
-            limit=limit,
-            engine_name=engine.name if engine else None,
-            engine_version=engine.version if engine else None,
-            access_control=access_control,
-            use_materialized=use_materialized,
-            query_parameters=query_parameters,
-        )
-        return translated_sql
-
-    # For all other nodes, build the node query
-    node = await Node.get_by_name(session, node_name, raise_if_not_exists=True)  # type: ignore
-    if node.type == NodeType.METRIC:
-        translated_sql, engine, _ = await build_sql_for_multiple_metrics(
-            session,
-            [node_name],
-            dimensions,
-            filters,
-            orderby,
-            limit,
-            engine.name if engine else None,
-            engine.version if engine else None,
-            access_control=access_control,
-            ignore_errors=ignore_errors,
-            use_materialized=use_materialized,
-            query_parameters=query_parameters,
-        )
-        query = translated_sql.sql
-        columns = translated_sql.columns
-    else:
-        query_ast = await get_query(
-            session=session,
-            node_name=node_name,
-            dimensions=dimensions,
-            filters=filters,
-            orderby=orderby,
-            limit=limit,
-            engine=engine,
-            access_control=access_control,
-            use_materialized=use_materialized,
-            query_parameters=query_parameters,
-            ignore_errors=ignore_errors,
-        )
-        columns = [
-            assemble_column_metadata(col, use_semantic_metadata=True)  # type: ignore
-            for col in query_ast.select.projection
-        ]
-        query = str(query_ast)
-
-    return TranslatedSQL.create(
-        sql=query,
-        columns=columns,
-        dialect=engine.dialect if engine else None,
-    )
-
-
-async def get_node_sql(
-    node_name: str,
-    dimensions: List[str] = [],
-    filters: List[str] = [],
-    orderby: List[str] = [],
-    limit: Optional[int] = None,
-    *,
-    session: AsyncSession = Depends(get_session),
-    engine_name: Optional[str] = None,
-    engine_version: Optional[str] = None,
-    current_user: User,
-    validate_access: access.ValidateAccessFn,
-    background_tasks: BackgroundTasks,
-    ignore_errors: bool = True,
-    use_materialized: bool = True,
-    query_parameters: dict[str, Any] | None = None,
-    cache: Cache = None,
-) -> Tuple[TranslatedSQL, QueryRequest]:
-    """
-    Return SQL for a node.
-    """
-    dimensions = [dim for dim in dimensions if dim and dim != ""]
-    access_control = access.AccessControlStore(
-        validate_access=validate_access,
-        user=UserOutput.from_orm(current_user),
-        base_verb=access.ResourceRequestVerb.READ,
-    )
-
-    engine = (
-        await get_engine(session, engine_name, engine_version)  # type: ignore
-        if engine_name
-        else None
-    )
-    validate_orderby(orderby, [node_name], dimensions)
-
-    query_request = await build_node_sql(
-        node_name=node_name,
-        dimensions=dimensions,
-        filters=filters,
-        orderby=orderby,
-        limit=limit,
-        session=session,
-        engine=engine,  # type: ignore
-        access_control=access_control,
-        ignore_errors=ignore_errors,
-        use_materialized=use_materialized,
-        query_parameters=query_parameters,
-    )
-    return (
-        TranslatedSQL.create(
-            sql=query_request.sql,
-            columns=query_request.columns,
-            dialect=engine.dialect if engine else None,
-        ),
-        query_request,
-    )
-
-
 @router.get(
     "/sql/{node_name}/",
     response_model=TranslatedSQL,
@@ -349,17 +183,12 @@ async def get_sql_for_metrics(
     ignore_errors: Optional[bool] = True,
     use_materialized: Optional[bool] = True,
     background_tasks: BackgroundTasks,
+    cache: Cache = Depends(get_cache),
+    request: Request,
 ) -> TranslatedSQL:
     """
     Return SQL for a set of metrics with dimensions and filters
     """
-
-    access_control = access.AccessControlStore(
-        validate_access=validate_access,
-        user=current_user,
-        base_verb=access.ResourceRequestVerb.READ,
-    )
-
     # make sure all metrics exist and have correct node type
     nodes = [
         await Node.get_by_name(session, node, raise_if_not_exists=True)
@@ -374,108 +203,25 @@ async def get_sql_for_metrics(
             http_status_code=HTTPStatus.UNPROCESSABLE_ENTITY,
         )
 
-    if (
-        query_request := await QueryRequest.get_query_request(
-            session,
+    query_cache_manager = QueryCacheManager(
+        cache=cache,
+        query_type=QueryBuildType.METRICS,
+    )
+    return await query_cache_manager.get_or_load(
+        background_tasks,
+        request,
+        QueryRequestParams(
             nodes=metrics,
             dimensions=dimensions,
             filters=filters,
-            orderby=orderby,
             limit=limit,
+            orderby=orderby,
+            query_params=query_params,
             engine_name=engine_name,
             engine_version=engine_version,
-            query_type=QueryBuildType.METRICS,
-        )
-    ) and not query_params:
-        # Update the node SQL in a background task to keep it up-to-date
-        background_tasks.add_task(  # pragma: no cover
-            build_and_save_sql_for_metrics,
-            session=session,
-            metrics=metrics,
-            dimensions=dimensions,
-            filters=filters,
-            orderby=orderby,
-            limit=limit,
-            engine_name=engine_name,
-            engine_version=engine_version,
-            access_control=access_control,
-            ignore_errors=ignore_errors,
             use_materialized=use_materialized,
-            query_parameters=json.loads(query_params),
-            save=False,
-        )
-        engine = (  # pragma: no cover
-            await get_engine(session, engine_name, engine_version)  # type: ignore
-            if engine_name
-            else None
-        )
-        return TranslatedSQL.create(  # pragma: no cover
-            sql=query_request.query,
-            columns=query_request.columns,
-            dialect=engine.dialect if engine else None,
-        )
-
-    return await build_and_save_sql_for_metrics(
-        session,
-        metrics,
-        dimensions,
-        filters,
-        orderby,
-        limit,
-        engine_name,
-        engine_version,
-        access_control,
-        ignore_errors=ignore_errors,  # type: ignore
-        use_materialized=use_materialized,  # type: ignore
-        query_parameters=json.loads(query_params),
-        save=False,
+            ignore_errors=ignore_errors,
+            current_user=current_user,
+            validate_access=validate_access,
+        ),
     )
-
-
-async def build_and_save_sql_for_metrics(
-    session: AsyncSession,
-    metrics: List[str],
-    dimensions: List[str],
-    filters: List[str] = None,
-    orderby: List[str] = None,
-    limit: Optional[int] = None,
-    engine_name: Optional[str] = None,
-    engine_version: Optional[str] = None,
-    access_control: Optional[access.AccessControlStore] = None,
-    ignore_errors: bool = True,
-    use_materialized: bool = True,
-    query_parameters: dict[str, Any] | None = None,
-    save: bool = True,
-):
-    """
-    Builds and saves SQL for metrics.
-    """
-    translated_sql, _, _ = await build_sql_for_multiple_metrics(
-        session,
-        metrics,
-        dimensions,
-        filters,
-        orderby,
-        limit,
-        engine_name,
-        engine_version,
-        access_control,
-        ignore_errors=ignore_errors,  # type: ignore
-        use_materialized=use_materialized,  # type: ignore
-        query_parameters=query_parameters,
-    )
-    if save:
-        await QueryRequest.save_query_request(  # pragma: no cover
-            session=session,
-            nodes=metrics,
-            dimensions=dimensions,
-            filters=filters,  # type: ignore
-            orderby=orderby,  # type: ignore
-            limit=limit,
-            engine_name=engine_name,
-            engine_version=engine_version,
-            query_type=QueryBuildType.METRICS,
-            query=translated_sql.sql,
-            columns=[col.dict() for col in translated_sql.columns],  # type: ignore
-        )
-    return translated_sql
