+++ datajunction-server/datajunction_server/internal/caching/query_cache_manager.py
@@ -5,18 +5,20 @@ import logging
 from typing import Any, OrderedDict
 from fastapi import Request
 from datajunction_server.internal.caching.cache_manager import RefreshAheadCacheManager
-
+from sqlalchemy.ext.asyncio import AsyncSession
 from datajunction_server.internal.caching.interface import Cache
 from datajunction_server.database.queryrequest import (
     QueryRequestKey,
     QueryBuildType,
     VersionedQueryKey,
 )
+from datajunction_server.internal.sql import build_sql_for_multiple_metrics
 from datajunction_server.database.user import User
 from datajunction_server.models import access
 from datajunction_server.models.sql import GeneratedSQL
 from datajunction_server.utils import session_context, get_settings
-from datajunction_server.construction.build_v2 import get_measures_query
+from datajunction_server.internal.sql import get_measures_query
+from datajunction_server.internal.sql import build_node_sql
 from datajunction_server.internal.engines import get_engine
 from datajunction_server.models.metric import TranslatedSQL
 
@@ -76,68 +78,43 @@ class QueryCacheManager(RefreshAheadCacheManager):
         self,
         request: Request,
         params: QueryRequestParams,
-    ) -> list[GeneratedSQL]:
+    ) -> list[GeneratedSQL] | TranslatedSQL:
         """
         The fallback function to call if the cache is not hit. This should be overridden
         in subclasses.
         """
         params = deepcopy(params)
         async with session_context(request) as session:
-            nodes = list(OrderedDict.fromkeys(params.nodes))
-            if self.query_type == QueryBuildType.MEASURES:
-                measures_query = await get_measures_query(
-                    session=session,
-                    metrics=nodes,
-                    dimensions=params.dimensions or [],
-                    filters=params.filters or [],
-                    orderby=params.orderby or [],
-                    engine_name=params.engine_name,
-                    engine_version=params.engine_version,
-                    current_user=params.current_user,
-                    validate_access=params.validate_access,
-                    include_all_columns=params.include_all_columns,
-                    use_materialized=params.use_materialized,
-                    preagg_requested=params.preaggregate,
-                    query_parameters=json.loads(params.query_params)
-                    if params.query_params
-                    else {},
-                )
-                return measures_query
-            if self.query_type == QueryBuildType.NODE:
-                from datajunction_server.api.sql import build_node_sql
-
-                engine = (
-                    await get_engine(session, params.engine_name, params.engine_version)  # type: ignore
-                    if params.engine_name
-                    else None
-                )
-                built_sql = await build_node_sql(
-                    node_name=nodes[0],
-                    dimensions=[
-                        dim for dim in (params.dimensions or []) if dim and dim != ""
-                    ],
-                    filters=params.filters or [],
-                    orderby=params.orderby or [],
-                    limit=params.limit,
-                    session=session,
-                    engine=engine,  # type: ignore
-                    ignore_errors=params.ignore_errors,
-                    use_materialized=params.use_materialized,
-                    query_parameters=json.loads(params.query_params)
-                    if params.query_params
-                    else {},
-                    access_control=access.AccessControlStore(
-                        validate_access=params.validate_access,
-                        user=params.current_user,
-                        base_verb=access.ResourceRequestVerb.READ,
-                    ),
-                )
-                return TranslatedSQL.create(
-                    sql=built_sql.sql,
-                    columns=built_sql.columns,
-                    dialect=built_sql.dialect,
-                )
-            raise ValueError(f"Unknown query type: {self.query_type}")
+            params.nodes = list(OrderedDict.fromkeys(params.nodes))
+            query_parameters = (
+                json.loads(params.query_params) if params.query_params else {}
+            )
+            access_control_store = access.AccessControlStore(
+                validate_access=params.validate_access,
+                user=params.current_user,
+                base_verb=access.ResourceRequestVerb.READ,
+            )
+            match self.query_type:
+                case QueryBuildType.MEASURES:
+                    return await self._build_measures_query(
+                        session,
+                        params,
+                        query_parameters,
+                    )
+                case QueryBuildType.NODE:
+                    return await self._build_node_query(
+                        session,
+                        params,
+                        query_parameters,
+                        access_control_store,
+                    )
+                case QueryBuildType.METRICS:
+                    return await self._build_metrics_query(
+                        session,
+                        params,
+                        query_parameters,
+                        access_control_store,
+                    )
 
     async def build_cache_key(
         self,
@@ -168,3 +145,83 @@ class QueryCacheManager(RefreshAheadCacheManager):
                 other_args=params.other_args or {},
             )
             return await super().build_cache_key(request, asdict(query_request))
+
+    async def _build_measures_query(
+        self,
+        session: AsyncSession,
+        params: QueryRequestParams,
+        query_parameters: dict[str, Any],
+    ) -> list[GeneratedSQL]:
+        return await get_measures_query(
+            session=session,
+            metrics=params.nodes,
+            dimensions=params.dimensions or [],
+            filters=params.filters or [],
+            orderby=params.orderby or [],
+            engine_name=params.engine_name,
+            engine_version=params.engine_version,
+            current_user=params.current_user,
+            validate_access=params.validate_access,
+            include_all_columns=params.include_all_columns,
+            use_materialized=params.use_materialized,
+            preagg_requested=params.preaggregate,
+            query_parameters=query_parameters,
+        )
+
+    async def _build_node_query(
+        self,
+        session: AsyncSession,
+        params: QueryRequestParams,
+        query_parameters: dict[str, Any],
+        access_control_store: access.AccessControlStore,
+    ) -> TranslatedSQL:
+        engine = (
+            await get_engine(session, params.engine_name, params.engine_version)  # type: ignore
+            if params.engine_name
+            else None
+        )
+        built_sql = await build_node_sql(
+            node_name=params.nodes[0],
+            dimensions=[dim for dim in (params.dimensions or []) if dim and dim != ""],
+            filters=params.filters or [],
+            orderby=params.orderby or [],
+            limit=params.limit,
+            session=session,
+            engine=engine,  # type: ignore
+            ignore_errors=params.ignore_errors,
+            use_materialized=params.use_materialized,
+            query_parameters=query_parameters,
+            access_control=access_control_store,
+        )
+        return TranslatedSQL.create(
+            sql=built_sql.sql,
+            columns=built_sql.columns,
+            dialect=built_sql.dialect,
+        )
+
+    async def _build_metrics_query(
+        self,
+        session: AsyncSession,
+        params: QueryRequestParams,
+        query_parameters: dict[str, Any],
+        access_control_store: access.AccessControlStore,
+    ) -> TranslatedSQL:
+        built_sql, _, _ = await build_sql_for_multiple_metrics(
+            session=session,
+            metrics=params.nodes,
+            dimensions=[dim for dim in (params.dimensions or []) if dim and dim != ""],
+            filters=params.filters or [],
+            orderby=params.orderby,
+            limit=params.limit,
+            engine_name=params.engine_name,
+            engine_version=params.engine_version,
+            access_control=access_control_store,
+            ignore_errors=params.ignore_errors,  # type: ignore
+            query_parameters=query_parameters,
+            use_materialized=params.use_materialized,  # type: ignore
+        )
+        return TranslatedSQL.create(
+            sql=built_sql.sql,
+            columns=built_sql.columns,
+            dialect=built_sql.dialect,
+        )
